IOS-1613 Move fit detents to file
